Select: migrate dropdowns to select component
