+++ app/components/forms/select/component.stories.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Story } from '@storybook/react/types-6-0';
-import Select, { SelectProps } from './component';
+import Select from './component';
+import { SelectProps } from './types';
 
 export default {
   title: 'Components/Forms/Select',
@@ -13,25 +14,51 @@ export default {
         options: ['dark', 'light'],
       },
     },
-    status: {
+    state: {
       control: {
         type: 'select',
-        options: ['none', 'valid', 'error', 'disabled'],
+        options: ['valid', 'error', 'none'],
+      },
+    },
+    onSelect: {
+      table: {
+        disable: true,
+      },
+    },
+    initialValues: {
+      table: {
+        disable: true,
       },
     },
   },
 };
 
-const Template: Story<SelectProps> = (args) => <Select {...args} />;
+const Template: Story<SelectProps> = (args) => (
+  <div className="relative">
+    <Select {...args} />
+  </div>
+);
 
 export const Default = Template.bind({});
 Default.args = {
   theme: 'dark',
-  placeholder: 'Select option',
+  size: 'base',
+  status: 'none',
+  prefix: 'FILTER BY:',
+  placeholder: 'Select Scenario',
   options: [
-    { label: 'Option 1', value: 'option-1' },
-    { label: 'Option 2', value: 'option-2' },
-    { label: 'Option 3', value: 'option-3', disabled: true },
-    { label: 'Option 4', value: 'option-4' },
+    { label: 'Scenario 1', value: 'scenario-1' },
+    { label: 'Scenario 2', value: 'scenario-2' },
+    { label: 'Scenario 3', value: 'scenario-3', disabled: true },
+    { label: 'Scenario 4', value: 'scenario-4' },
   ],
+  initialSelected: ['scenario-1', 'scenario-2', 'scenario-4'],
+  disabled: false,
+  multiple: true,
+  searchable: false,
+  clearSelectionActive: true,
+  clearSelectionLabel: 'Clear Selection',
+  batchSelectionActive: true,
+  batchSelectionLabel: 'Select all',
+  onChange: (option) => console.info(option),
 };
