+++ app/components/forms/select/multi/component.tsx
@@ -4,20 +4,20 @@ import cx from 'classnames';
 
 // Downshift;
 import { useSelect, useMultipleSelection } from 'downshift';
-import Toggle from 'components/dropdowns/toggle';
-import Menu from 'components/dropdowns/menu';
+import Toggle from 'components/forms/select/toggle';
+import Menu from 'components/forms/select/menu';
 import Checkbox from 'components/forms/checkbox';
 
 // Popper
 import { usePopper } from 'react-popper';
 import {
   flipModifier, hideModifier, sameWidthModifier, offsetModifier,
-} from 'components/dropdowns/constants/popper-modifiers';
-import THEME from 'components/dropdowns/constants/theme';
+} from 'components/forms/select/constants/popper-modifiers';
+import THEME from 'components/forms/select/constants/theme';
 
-import { DropdownProps, DropdownOptionProps } from 'components/dropdowns/types';
+import { SelectProps, SelectOptionProps } from 'components/forms/select/types';
 
-export const MultiDropdown: React.FC<DropdownProps> = ({
+export const MultiSelect: React.FC<SelectProps> = ({
   theme = 'dark',
   size = 'base',
   status,
@@ -34,7 +34,7 @@ export const MultiDropdown: React.FC<DropdownProps> = ({
   onSelect,
   onFocus,
   onBlur,
-}: DropdownProps) => {
+}: SelectProps) => {
   const triggerRef = useRef();
   const menuRef = useRef();
 
@@ -70,7 +70,7 @@ export const MultiDropdown: React.FC<DropdownProps> = ({
     return getOptions.filter((o) => initialValues.includes(`${o.value}`));
   }, [getOptions, initialValues]);
 
-  const isSelected = (selected: DropdownOptionProps, selectedItms: DropdownOptionProps[]) => (
+  const isSelected = (selected: SelectOptionProps, selectedItms: SelectOptionProps[]) => (
     selectedItms.some((i) => i.value === selected.value)
   );
 
@@ -298,4 +298,4 @@ export const MultiDropdown: React.FC<DropdownProps> = ({
   );
 };
 
-export default MultiDropdown;
+export default MultiSelect;
